refactor(api): extract project work projection
